Add user registration functionality
Added functionality that allows the creation of new users. A new UserFormDto.java file was added to house the user registration form data, and the relevant HTML templates were updated to provide a UI for user creation. The HomeController.java file was also updated to handle the new user creation requests.