Refactor methods and template fragments to enhance usability
Renamed method 'addTodoItem' to 'addUserTask' for better clarity of its functionality in HomeController.java. Also, erased commented code since it was no longer necessary. Modified 'userFragment.html' by removing 'hx-push-url' attribute from the form element, as it was causing unnecessary page reload. In 'tasksFragment.html', modified form with 'hx:post' and input field for better structuring and refactored table layout for improved UI. 'taskFragment.html' was updated to include task in a table row for better readability. The changes are made to make the code cleaner and more understandable, and also to enhance the UI/UX.